Fix crash: messages with incorrect capitalization
Resolves jethornton/JT-Logger#12

Sometimes, channel names coming in from a message
may differ from the channel name we are monitoring, only
in terms of capitalization. For example: if we are monitoring
'#channel', sometimes, a person may join the channel using the
command
    /join #CHANNEL
or
    /join #Channel

In either case, the server may not care - and allow the user to
join #channel using the above commands. (At least, that is the
case with ngircd)

For logging purposes - we only want ONE file for logging the events.

Also, we should not crash with something like:

File "/srv/irclogger3/JT-Logger/jt-logger.py", line 361, in create_log
    with open(path, 'w') as log:
IOError: [Errno 2] No such file or directory: '/srv/www/irc_log_dir/#Channel/2019-06-28.html'